requirements.txt
Django>=5.0.0
Pillow>=10.0.0
qrcode>=7.4.2
django-crispy-forms>=2.0
crispy-tailwind>=0.5.0
django-filter>=23.5
django-tables2>=2.6.0
django-import-export>=3.3.3
django-widget-tweaks>=1.5.0
django-allauth>=0.58.2
django-extensions>=3.2.3
django-debug-toolbar>=4.2.0
whitenoise>=6.6.0
gunicorn>=21.2.0
python-dotenv>=1.0.0